autonomous_quant_trading_bot/requirements.txt
numpy>=1.24
scipy>=1.11
pandas>=2.0
statsmodels>=0.14
vectorbt>=0.26
MetaTrader5>=5.0.45
ccxt>=4.0
PyYAML>=6.0
scikit-learn>=1.3
matplotlib>=3.7
rich>=13.0
